% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcite-package.R
\docType{package}
\name{calcite-package}
\alias{calcite}
\alias{calcite-package}
\title{calcite: Bindings to the Calcite Design System 'JavaScript' Component Library}
\description{
Provides access to the 'Calcite Design System' 'javascript' components via integration with the 'htmltools' and 'shiny' packages. Pre-built and interactive components can be used to generate either static html or interactive web applications. Learn more about the 'Calcite Design System' at \url{https://developers.arcgis.com/calcite-design-system/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.esri.com/calcite}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
