% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_action}
\alias{calcite_action}
\title{Calcite Action}
\usage{
calcite_action(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-action>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   active \tab active \tab When true, the component is highlighted. \tab boolean \tab false \cr
   alignment \tab alignment \tab Specifies the horizontal alignment of button elements with text content. \tab "center" | "end" | "start" \tab undefined \cr
   appearance \tab appearance \tab Specifies the appearance of the component. \tab "solid" | "transparent" \tab "solid" \cr
   compact \tab compact \tab DEPRECATED No longer necessary.When true, the side padding of the component is reduced. \tab boolean \tab false \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   icon \tab icon \tab Specifies an icon to display. \tab string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab When true, the icon will be flipped when the element direction is right-to-left ("rtl"). \tab boolean \tab false \cr
   indicator \tab indicator \tab When true, displays a visual indicator. \tab boolean \tab false \cr
   label \tab label \tab Specifies the label of the component. If no label is provided, the label inherits what’s provided for the text prop. \tab string \tab undefined \cr
   loading \tab loading \tab When true, a busy indicator is displayed. \tab boolean \tab false \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab ActionMessages \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   text (required) \tab text \tab Specifies text that accompanies the icon. \tab string \tab undefined \cr
   textEnabled \tab text-enabled \tab Indicates whether the text is displayed. \tab boolean \tab false \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding a calcite-icon. \cr
   "tooltip" \tab DEPRECATED Use the calcite-tooltip component instead. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/action/}{Official Documentation}
}
