% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_block_section}
\alias{calcite_block_section}
\title{Calcite Block Section}
\usage{
calcite_block_section(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-block-section>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/block-section/}{Official Documentation}
}
