% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_chip_group}
\alias{calcite_chip_group}
\title{Calcite Chip Group}
\usage{
calcite_chip_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-chip-group>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. Child calcite-chips inherit the component’s value. \tab "l" | "m" | "s" \tab "m" \cr
   selectedItems \tab – \tab Specifies the component’s selected items. \tab HTMLCalciteChipElement[] \tab [] \cr
   selectionMode \tab selection-mode \tab Specifies the selection mode of the component, where: "multiple" allows any number of selections, "single" allows only one selection, "single-persist" allows one selection and prevents de-selection, and "none" does not allow any selections. \tab "multiple" | "none" | "single" | "single-persist" \tab "none" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteChipGroupSelect \tab Fires when the component’s selection changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding one or more calcite-chips. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/chip-group/}{Official Documentation}
}
