% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_color_picker_hex_input}
\alias{calcite_color_picker_hex_input}
\title{Calcite Color Picker Hex Input}
\usage{
calcite_color_picker_hex_input(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-color-picker-hex-input>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/color-picker-hex-input/}{Official Documentation}
}
