% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_color_picker_swatch}
\alias{calcite_color_picker_swatch}
\title{Calcite Color Picker Swatch}
\usage{
calcite_color_picker_swatch(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-color-picker-swatch>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/color-picker-swatch/}{Official Documentation}
}
