% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_date_picker_month}
\alias{calcite_date_picker_month}
\title{Calcite Date Picker Month}
\usage{
calcite_date_picker_month(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-date-picker-month>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   activeDate \tab – \tab The currently active Date. \tab Date \tab new Date() \cr
   endDate \tab – \tab End date currently active. \tab Date \tab undefined \cr
   headingLevel \tab heading-level \tab Specifies the number at which section headings should start. \tab 1 | 2 | 3 | 4 | 5 | 6 \tab undefined \cr
   hoverRange \tab – \tab The range of dates currently being hovered. \tab HoverRange \tab undefined \cr
   max \tab – \tab Specifies the latest allowed date ("yyyy-mm-dd"). \tab Date \tab undefined \cr
   min \tab – \tab Specifies the earliest allowed date ("yyyy-mm-dd"). \tab Date \tab undefined \cr
   monthStyle \tab month-style \tab Specifies the monthStyle used by the component. \tab "abbreviated" | "wide" \tab undefined \cr
   range \tab range \tab When true, activates the component’s range mode which renders two calendars for selecting ranges of dates. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab undefined \cr
   selectedDate \tab – \tab Already selected date. \tab Date \tab undefined \cr
   startDate \tab – \tab Start date currently active. \tab Date \tab undefined \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/date-picker-month/}{Official Documentation}
}
