% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown.R
\name{calcite_dropdown}
\alias{calcite_dropdown}
\title{Create a Calcite dropdown Component}
\usage{
calcite_dropdown(
  ...,
  id = NULL,
  close_on_select_disabled = FALSE,
  disabled = FALSE,
  max_items = 0L,
  offset_distance = 0L,
  offset_skidding = 0L,
  open = FALSE,
  overlay_positioning = "absolute",
  placement = "bottom-start",
  scale = "s",
  trigger = NULL,
  type = "click",
  width = "s"
)
}
\arguments{
\item{...}{One or more \link{calcite_dropdown_item} components}

\item{id}{Component ID (required for Shiny reactivity)}

\item{scale}{Size of the component: "s" (small), "m" (medium), or "l" (large) (default: "m")}

\item{selection_mode}{Selection mode: "multiple" allows any number of selections, "single" allows only one selection, "single-persist" allows one selection and prevents de-selection (default: "multiple")}
}
\value{
An object of class \code{calcite_component}
}
\description{
Create a Calcite dropdown Component
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/dropdown/}{Official Documentation}
}
