% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_flow}
\alias{calcite_flow}
\title{Calcite Flow}
\usage{
calcite_flow(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-flow>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/flow/}{Official Documentation}
}
