% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_flow_item}
\alias{calcite_flow_item}
\title{Create a FlowItem component}
\usage{
calcite_flow_item(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a FlowItem component
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   beforeBack \tab NA \tab Specifies a function to run before the component is removed from its parent \code{calcite-flow}. \tab Check API reference \tab FALSE \cr
   beforeClose \tab NA \tab Specifies a function to run before the component closes. \tab Check API reference \tab FALSE \cr
   closable \tab closable \tab When \code{true}, displays a close button in the trailing side of the component's header. \tab boolean \tab TRUE \cr
   closed \tab closed \tab When \code{true}, hides the component. \tab boolean \tab TRUE \cr
   collapsed \tab collapsed \tab When \code{true}, hides the component's content area. \tab boolean \tab TRUE \cr
   collapsible \tab collapsible \tab When \code{true}, the component is collapsible. \tab boolean \tab TRUE \cr
   description \tab description \tab Specifies a the component's description. \tab string \tab FALSE \cr
   disabled \tab disabled \tab When \code{true}, prevents interaction and decreases the component's opacity. \tab boolean \tab TRUE \cr
   heading \tab heading \tab Specifies the component's heading text. \tab string \tab FALSE \cr
   headingLevel \tab heading-level \tab Specifies the heading level number of the component's \code{heading} for proper document structure, without affecting visual styling. \tab HeadingLevel \tab TRUE \cr
   icon \tab icon \tab Specifies an icon to display. \tab IconName \tab TRUE \cr
   iconFlipRtl \tab icon-flip-rtl \tab When \code{true} and the element direction is right-to-left (\code{"rtl"}), flips the component\code{s }icon`. \tab boolean \tab TRUE \cr
   loading \tab loading \tab When \code{true}, a busy indicator is displayed. \tab boolean \tab TRUE \cr
   menuOpen \tab menu-open \tab When \code{true}, the action menu items in the \code{header-menu-actions} slot are open. \tab boolean \tab TRUE \cr
   messageOverrides \tab NA \tab Overrides individual strings used by the component. \tab Check API reference \tab FALSE \cr
   overlayPositioning \tab overlay-positioning \tab Specifies the type of positioning to use for overlaid content, where:  \code{"absolute"} works for most cases - positioning the component inside of overflowing parent containers, which affects the container's layout, and  \code{"fixed"} is used to escape an overflowing parent container, or when the reference element's \code{position} CSS property is \code{"fixed"}. \tab OverlayPositioning \tab TRUE \cr
   scale \tab scale \tab Specifies the size of the component. \tab Scale \tab TRUE \cr
   selected \tab selected \tab When \code{true}, the component is displayed within a parent \code{calcite-flow}. \tab boolean \tab TRUE \cr
   showBackButton \tab show-back-button \tab When \code{true}, displays a back button in the component's header. \tab boolean \tab FALSE \cr
   topLayerDisabled \tab top-layer-disabled \tab When \code{true} and the component is \code{open}, disables top layer placement.  Only set this if you need complex z-index control or if top layer placement causes conflicts with third-party components. \tab boolean \tab TRUE \cr
}

}

\subsection{Events}{

The following events are observed by shiny:\tabular{ll}{
   Event \tab Description \cr
   calciteFlowItemBack \tab Fires when the component's back button is clicked. \cr
   calciteFlowItemClose \tab Fires when the component's close button is clicked. \cr
   calciteFlowItemCollapse \tab Fires when the component's content area is collapsed. \cr
   calciteFlowItemExpand \tab Fires when the component's content area is expanded. \cr
   calciteFlowItemScroll \tab Fires when the component's content is scrolled. \cr
   calciteFlowItemToggle \tab Fires when the component's collapse button is clicked. \cr
}

}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding custom content. \cr
   action-bar \tab A slot for adding a \code{calcite-action-bar} to the component. \cr
   alerts \tab A slot for adding \code{calcite-alert}s to the component. \cr
   content-top \tab A slot for adding content above the unnamed (default) slot and below the action-bar slot (if populated). \cr
   content-bottom \tab A slot for adding content below the unnamed (default) slot and above the footer slot (if populated) \cr
   header-actions-start \tab A slot for adding \code{calcite-action}s or content to the start side of the component's header. \cr
   header-actions-end \tab A slot for adding \code{calcite-action}s or content to the end side of the component's header. \cr
   header-content \tab A slot for adding custom content to the component's header. \cr
   header-menu-actions \tab A slot for adding an overflow menu with \code{calcite-action}s inside a \code{calcite-dropdown}. \cr
   fab \tab A slot for adding a \code{calcite-fab} (floating action button) to perform an action. \cr
   footer \tab A slot for adding custom content to the component's footer. Should not be used with the \code{footer-start} or \code{footer-end} slots. \cr
   footer-end \tab A slot for adding a trailing footer custom content. Should not be used with the \code{footer} slot. \cr
   footer-start \tab A slot for adding a leading footer custom content. Should not be used with the \code{footer} slot. \cr
}

}
}
\examples{
calcite_flow_item()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/flow-item/}{Official Documentation}
}
