% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_icon}
\alias{calcite_icon}
\title{Calcite Icon}
\usage{
calcite_icon(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-icon>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/icon/}{Official Documentation}
}
