% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_inline_editable}
\alias{calcite_inline_editable}
\title{Calcite Inline Editable}
\usage{
calcite_inline_editable(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-inline-editable>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/inline-editable/}{Official Documentation}
}
