% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_input}
\alias{calcite_input}
\title{Calcite Input}
\usage{
calcite_input(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-input>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   accept \tab accept \tab Specifies a comma separated list of unique file type specifiers for limiting accepted file types. This property only has an effect when type is “file”. Read the native attribute’s documentation on MDN for more info. \tab string \tab undefined \cr
   alignment \tab alignment \tab Specifies the text alignment of the component’s value. \tab "end" | "start" \tab "start" \cr
   autocomplete \tab autocomplete \tab Specifies the type of content to autocomplete, for use in forms. Read the native attribute’s documentation on MDN for more info. \tab string \tab undefined \cr
   autofocus \tab autofocus \tab Adds global prop, missing from Stencil’s HTMLElement type, see https://github.com/ionic-team/stencil/issues/5726 \tab boolean \tab undefined \cr
   clearable \tab clearable \tab When true, a clear button is displayed when the component has a value. The clear button shows by default for "search", "time", and "date" types, and will not display for the "textarea" type. \tab boolean \tab false \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   files \tab – \tab When type is "file", specifies the component’s selected files. \tab FileList \tab undefined \cr
   form \tab form \tab The id of the form that will be associated with the component. When not set, the component will be associated with its ancestor form element, if any. \tab string \tab undefined \cr
   groupSeparator \tab group-separator \tab When true, number values are displayed with a group separator corresponding to the language and country format. \tab boolean \tab false \cr
   icon \tab icon \tab When true, shows a default recommended icon. Alternatively, pass a Calcite UI Icon name to display a specific icon. \tab boolean | string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab When true, the icon will be flipped when the element direction is right-to-left ("rtl"). \tab boolean \tab false \cr
   label \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   loading \tab loading \tab When true, a busy indicator is displayed. \tab boolean \tab false \cr
   max \tab max \tab Specifies the maximum value for type “number”. \tab number \tab undefined \cr
   maxLength \tab max-length \tab Specifies the maximum length of text for the component’s value. \tab number \tab undefined \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab InputMessages \tab undefined \cr
   min \tab min \tab Specifies the minimum value for type="number". \tab number \tab undefined \cr
   minLength \tab min-length \tab Specifies the minimum length of text for the component’s value. \tab number \tab undefined \cr
   multiple \tab multiple \tab When true, the component can accept more than one value. This property only has an effect when type is “email” or “file”. Read the native attribute’s documentation on MDN for more info. \tab boolean \tab false \cr
   name \tab name \tab Specifies the name of the component. Required to pass the component’s value on form submission. \tab string \tab undefined \cr
   numberButtonType \tab number-button-type \tab Specifies the placement of the buttons for type="number". \tab "horizontal" | "none" | "vertical" \tab "vertical" \cr
   numberingSystem \tab numbering-system \tab Specifies the Unicode numeral system used by the component for localization. \tab "arab" | "arabext" | "latn" \tab undefined \cr
   pattern \tab pattern \tab Specifies a regex pattern the component’s value must match for validation. Read the native attribute’s documentation on MDN for more info. \tab string \tab undefined \cr
   placeholder \tab placeholder \tab Specifies placeholder text for the component. \tab string \tab undefined \cr
   prefixText \tab prefix-text \tab Adds text to the start of the component. \tab string \tab undefined \cr
   readOnly \tab read-only \tab When true, the component’s value can be read, but cannot be modified. \tab boolean \tab false \cr
   required \tab required \tab When true, the component must have a value in order for the form to submit. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   status \tab status \tab Specifies the status of the input field, which determines message and icons. \tab "idle" | "invalid" | "valid" \tab "idle" \cr
   step \tab step \tab Specifies the granularity the component’s value must adhere to. \tab "any" | number \tab undefined \cr
   suffixText \tab suffix-text \tab Adds text to the end of the component. \tab string \tab undefined \cr
   type \tab type \tab Specifies the component type. Note that the following types add type-specific icons by default: "date", "email", "password", "search", "tel", "time". \tab "color" | "date" | "datetime-local" | "email" | "file" | "image" | "month" | "number" | "password" | "search" | "tel" | "text" | "textarea" | "time" | "url" | "week" \tab "text" \cr
   validationIcon \tab validation-icon \tab Specifies the validation icon to display under the component. \tab boolean | string \tab undefined \cr
   validationMessage \tab validation-message \tab Specifies the validation message to display under the component. \tab string \tab undefined \cr
   validity \tab – \tab The current validation state of the component. \tab { valid: boolean; badInput: boolean; customError: boolean; patternMismatch: boolean; rangeOverflow: boolean; rangeUnderflow: boolean; stepMismatch: boolean; tooLong: boolean; tooShort: boolean; typeMismatch: boolean; valueMissing: boolean; } \tab {     valid: false,     badInput: false,     customError: false,     patternMismatch: false,     rangeOverflow: false,     rangeUnderflow: false,     stepMismatch: false,     tooLong: false,     tooShort: false,     typeMismatch: false,     valueMissing: false,   } \cr
   value \tab value \tab The component’s value. \tab string \tab "" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteInputChange \tab Fires each time a new value is typed and committed. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteInputInput \tab Fires each time a new value is typed. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   "action" \tab A slot for positioning a calcite-button next to the component. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/input/}{Official Documentation}
}
