% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{calcite_list}
\alias{calcite_list}
\title{Create a Calcite list Component}
\usage{
calcite_list(
  ...,
  id = NULL,
  drag_enabled = NULL,
  scale = NULL,
  selection_mode = NULL
)
}
\arguments{
\item{...}{One or more \code{\link[=calcite_list()_item]{calcite_list()_item()}} components}

\item{id}{Component ID (required for Shiny reactivity)}

\item{scale}{Size of the component: "s" (small), "m" (medium), or "l" (large) (default: "m")}

\item{selection_mode}{Selection mode: "multiple" allows any number of selections, "single" allows only one selection, "single-persist" allows one selection and prevents de-selection (default: "multiple")}
}
\value{
An object of class \code{calcite_component}
}
\description{
Create a Calcite list Component
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/list/}{Official Documentation}
}
