% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_loader}
\alias{calcite_loader}
\title{Calcite Loader}
\usage{
calcite_loader(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-loader>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   inline \tab inline \tab When true, displays smaller and appears to the left of the text. \tab boolean \tab false \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   text \tab text \tab Text that displays under the component’s indicator. \tab string \tab "" \cr
   type \tab type \tab Specifies the component type. Use "indeterminate" if finding actual progress value is impossible. Otherwise, use "determinate" to have the value indicate the progress or "determinate-value" to have the value label displayed along the progress. \tab "determinate" | "determinate-value" | "indeterminate" \tab "indeterminate" \cr
   value \tab value \tab The component’s value. Valid only for "determinate" indicators. Percent complete of 100. \tab number \tab 0 \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/loader/}{Official Documentation}
}
