% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_menu}
\alias{calcite_menu}
\title{Calcite Menu}
\usage{
calcite_menu(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-menu>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   layout \tab layout \tab Specifies the layout of the component. \tab "horizontal" | "vertical" \tab "horizontal" \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab MenuMessages \tab undefined \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/menu/}{Official Documentation}
}
