% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_menu_item}
\alias{calcite_menu_item}
\title{Calcite Menu Item}
\usage{
calcite_menu_item(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-menu-item>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   active \tab active \tab When true, the component is highlighted. \tab boolean \tab undefined \cr
   breadcrumb \tab breadcrumb \tab When true, the component displays a breadcrumb trail for use as a navigational aid. \tab boolean \tab undefined \cr
   href \tab href \tab Specifies the URL destination of the component, which can be set as an absolute or relative path. \tab string \tab undefined \cr
   iconEnd \tab icon-end \tab Specifies an icon to display at the end of the component. \tab string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab Displays the iconStart and/or iconEnd as flipped when the element direction is right-to-left ("rtl"). \tab "both" | "end" | "start" \tab undefined \cr
   iconStart \tab icon-start \tab Specifies an icon to display at the start of the component. \tab string \tab undefined \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab MenuItemMessages \tab undefined \cr
   open \tab open \tab When true, the component will display any slotted calcite-menu-item in an open overflow menu. \tab boolean \tab false \cr
   rel \tab rel \tab Defines the relationship between the href value and the current document. \tab string \tab undefined \cr
   target \tab target \tab Specifies where to open the linked document defined in the href property. \tab string \tab undefined \cr
   text \tab text \tab Specifies the text to display. \tab string \tab undefined \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteMenuItemSelect \tab Emits when the component is selected. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   "submenu-item" \tab A slot for adding calcite-menu-items in a submenu. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/menu-item/}{Official Documentation}
}
