% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_modal}
\alias{calcite_modal}
\title{Calcite Modal}
\usage{
calcite_modal(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-modal>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   beforeClose \tab – \tab Passes a function to run before the component closes. \tab (el: HTMLCalciteModalElement) => Promise\if{html}{\out{<void>}} \tab undefined \cr
   closeButtonDisabled \tab close-button-disabled \tab When true, disables the component’s close button. \tab boolean \tab false \cr
   docked \tab docked \tab When true, prevents the component from expanding to the entire screen on mobile devices. \tab boolean \tab undefined \cr
   escapeDisabled \tab escape-disabled \tab When true, disables the default close on escape behavior. \tab boolean \tab false \cr
   focusTrapDisabled \tab focus-trap-disabled \tab When true, prevents focus trapping. \tab boolean \tab false \cr
   fullscreen \tab fullscreen \tab Sets the component to always be fullscreen. Overrides widthScale and --calcite-modal-width / --calcite-modal-height. \tab boolean \tab undefined \cr
   kind \tab kind \tab Specifies the kind of the component, which will apply to top border. \tab "brand" | "danger" | "info" | "success" | "warning" \tab undefined \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab ModalMessages \tab undefined \cr
   open \tab open \tab When true, displays and positions the component. \tab boolean \tab false \cr
   outsideCloseDisabled \tab outside-close-disabled \tab When true, disables the closing of the component when clicked outside. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   widthScale \tab width-scale \tab Specifies the width of the component. \tab "l" | "m" | "s" \tab "m" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteModalBeforeClose \tab Fires when the component is requested to be closed and before the closing transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteModalBeforeOpen \tab Fires when the component is added to the DOM but not rendered, and before the opening transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteModalClose \tab Fires when the component is closed and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteModalOpen \tab Fires when the component is open and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   "back" \tab A slot for adding a back button. \cr
   "content" \tab A slot for adding the component’s content. \cr
   "content-bottom" \tab A slot for adding content to the component’s sticky footer, where content remains at the bottom of the component when scrolling up and down. \cr
   "content-top" \tab A slot for adding content to the component’s sticky header, where content remains at the top of the component when scrolling up and down. \cr
   "header" \tab A slot for adding header text. \cr
   "primary" \tab A slot for adding a primary button. \cr
   "secondary" \tab A slot for adding a secondary button. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/modal/}{Official Documentation}
}
