% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_navigation_user}
\alias{calcite_navigation_user}
\title{Create a NavigationUser component}
\usage{
calcite_navigation_user(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a NavigationUser component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   active \tab active \tab When \code{true}, the component is highlighted. \tab boolean \tab TRUE \cr
   fullName \tab full-name \tab Specifies the full name of the user. \tab string \tab FALSE \cr
   label \tab label \tab Describes the appearance of the avatar. If no label is provided, context will not be provided to assistive technologies. \tab string \tab FALSE \cr
   textDisabled \tab text-disabled \tab When \code{true}, hides the \code{fullName} and \code{username} contents. \tab boolean \tab TRUE \cr
   thumbnail \tab thumbnail \tab Specifies the \code{src} to an image (remember to add a token if the user is private). \tab string \tab FALSE \cr
   userId \tab user-id \tab Specifies the unique id of the user. \tab string \tab FALSE \cr
   username \tab username \tab Specifies the username of the user. \tab string \tab FALSE \cr
}
}
\examples{
calcite_navigation_user()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/navigation-user/}{Official Documentation}
}
