% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_popover}
\alias{calcite_popover}
\title{Create a Popover component}
\usage{
calcite_popover(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a Popover component
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   autoClose \tab auto-close \tab When \code{true}, clicking outside of the component automatically closes open \code{calcite-popover}s. \tab boolean \tab TRUE \cr
   closable \tab closable \tab When \code{true}, displays a close button in the component. \tab boolean \tab TRUE \cr
   flipDisabled \tab flip-disabled \tab When \code{true}, prevents flipping the component's placement when overlapping its \code{referenceElement}. \tab boolean \tab TRUE \cr
   flipPlacements \tab NA \tab Specifies the component's fallback \code{placement} for slotted content when it's initial or specified \code{placement} has insufficient space available. \tab Check API reference \tab FALSE \cr
   focusTrapDisabled \tab focus-trap-disabled \tab When \code{true}, prevents focus trapping. \tab boolean \tab TRUE \cr
   focusTrapOptions \tab NA \tab Specifies custom focus trap configuration on the component, where  \verb{"allowOutsideClick}" allows outside clicks, \code{"initialFocus"} enables initial focus, \code{"returnFocusOnDeactivate"} returns focus when not active, \code{"extraContainers"} specifies additional focusable elements external to the trap, such as 3rd-party components appending elements to the document body, and \code{"setReturnFocus"} customizes the element to which focus is returned when the trap is deactivated. Return \code{false} to prevent focus return, or \code{undefined} to use the default behavior (returning focus to the element focused before activation). \tab Check API reference \tab FALSE \cr
   heading \tab heading \tab Specifies the component's heading text. \tab string \tab FALSE \cr
   headingLevel \tab heading-level \tab Specifies the heading level number of the component's \code{heading} for proper document structure, without affecting visual styling. \tab HeadingLevel \tab TRUE \cr
   label \tab label \tab Specifies an accessible label for the component. \tab string \tab FALSE \cr
   messageOverrides \tab NA \tab Overrides individual strings used by the component. \tab Check API reference \tab FALSE \cr
   offsetDistance \tab offset-distance \tab Specifies the distance to position the component away from the \code{referenceElement}. \tab number \tab TRUE \cr
   offsetSkidding \tab offset-skidding \tab Specifies the distance to position the component along the \code{referenceElement}. \tab number \tab TRUE \cr
   open \tab open \tab When \code{true}, displays and positions the component. \tab boolean \tab TRUE \cr
   overlayPositioning \tab overlay-positioning \tab Specifies the type of positioning to use for overlaid content, where:  \code{"absolute"} works for most cases - positioning the component inside of overflowing parent containers, which affects the container's layout, and  \code{"fixed"} is used to escape an overflowing parent container, or when the reference element's \code{position} CSS property is \code{"fixed"}. \tab OverlayPositioning \tab TRUE \cr
   placement \tab placement \tab Determines where the component will be positioned relative to the \code{referenceElement}. \tab LogicalPlacement \tab TRUE \cr
   pointerDisabled \tab pointer-disabled \tab When \code{true}, removes the caret pointer. \tab boolean \tab TRUE \cr
   referenceElement \tab reference-element \tab The \code{referenceElement} is used to position the component according to its \code{placement} value.  Setting the value to an \code{HTMLElement} is preferred so the component does not need to query the DOM.  However, a string \code{id} of the reference element can also be used.  The component should not be placed within its own \code{referenceElement} to avoid unintended behavior. \tab ReferenceElement | string \tab FALSE \cr
   scale \tab scale \tab Specifies the size of the component. \tab Scale \tab TRUE \cr
   topLayerDisabled \tab top-layer-disabled \tab When \code{true} and the component is \code{open}, disables top layer placement.  Only set this if you need complex z-index control or if top layer placement causes conflicts with third-party components. \tab boolean \tab TRUE \cr
   triggerDisabled \tab trigger-disabled \tab When \code{true}, disables automatically toggling the component when its \code{referenceElement} has been triggered.  This property can be set to \code{true} to manage when the component is open. \tab boolean \tab TRUE \cr
}

}

\subsection{Events}{

The following events are observed by shiny:\tabular{ll}{
   Event \tab Description \cr
   calcitePopoverBeforeClose \tab Fires when the component is requested to be closed and before the closing transition begins. \cr
   calcitePopoverBeforeOpen \tab Fires when the component is added to the DOM but not rendered, and before the opening transition begins. \cr
   calcitePopoverClose \tab Fires when the component is closed and animation is complete. \cr
   calcitePopoverOpen \tab Fires when the component is open and animation is complete. \cr
}

}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding custom content. \cr
}

}
}
\examples{
calcite_popover()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/popover/}{Official Documentation}
}
