% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_progress}
\alias{calcite_progress}
\title{Calcite Progress}
\usage{
calcite_progress(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-progress>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   label \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   reversed \tab reversed \tab When true and for "indeterminate" progress bars, reverses the animation direction. \tab boolean \tab false \cr
   text \tab text \tab Text that displays under the component’s indicator. \tab string \tab undefined \cr
   type \tab type \tab Specifies the component type. Use "indeterminate" if finding actual progress value is impossible. \tab "determinate" | "indeterminate" \tab "determinate" \cr
   value \tab value \tab When type is "determinate", the component’s progress value with a range of 0.0 - 1.0. \tab number \tab 0 \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/progress/}{Official Documentation}
}
