% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R, R/segmented-control.R
\name{calcite_segmented_control_item}
\alias{calcite_segmented_control_item}
\title{Create a SegmentedControlItem component}
\usage{
calcite_segmented_control_item(
  value,
  label = value,
  checked = NULL,
  icon_start = NULL,
  icon_end = NULL,
  icon_flip_rtl = NULL,
  ...
)

calcite_segmented_control_item(
  value,
  label = value,
  checked = NULL,
  icon_start = NULL,
  icon_end = NULL,
  icon_flip_rtl = NULL,
  ...
)
}
\arguments{
\item{value}{The value this item represents (required)}

\item{checked}{Whether this item is initially selected (default: FALSE)}

\item{icon_start}{Icon to display at the start of the item}

\item{icon_end}{Icon to display at the end of the item}

\item{icon_flip_rtl}{Whether to flip the icon in RTL languages (default: FALSE)}

\item{...}{Additional content or attributes (text label if provided as unnamed argument)}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}

An object of class \code{calcite_component}
}
\description{
Creates an individual item within a segmented control. Must be used as a child
of \code{calcite_segmented_control()}.
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   checked \tab checked \tab When \code{true}, the component is checked. \tab boolean \tab TRUE \cr
   iconEnd \tab icon-end \tab Specifies an icon to display at the end of the component. \tab string \tab TRUE \cr
   iconFlipRtl \tab icon-flip-rtl \tab When \code{true}, the icon will be flipped when the element direction is right-to-left (\code{"rtl"}). \tab boolean \tab TRUE \cr
   iconStart \tab icon-start \tab Specifies an icon to display at the start of the component. \tab string \tab TRUE \cr
   value \tab value \tab The component's value. \tab any \tab FALSE \cr
}

}
}
\examples{
calcite_segmented_control_item()
# Item with text label
calcite_segmented_control_item(value = "option1")

# Item with icon
calcite_segmented_control_item(
  value = "list",
  icon_start = "list"
)

# Initially selected item
calcite_segmented_control_item(
  value = "default",
  checked = TRUE
)
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/segmented-control-item/}{Official Documentation}

\href{https://developers.arcgis.com/calcite-design-system/components/segmented-control-item/}{Official Documentation}
}
