% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_sheet}
\alias{calcite_sheet}
\title{Create a Sheet component}
\usage{
calcite_sheet(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a Sheet component
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   beforeClose \tab NA \tab Passes a function to run before the component closes. \tab Check API reference \tab FALSE \cr
   displayMode \tab display-mode \tab Specifies the display mode - \code{"float"} separates content from main layout, and \code{"overlay"} displays on top of center content. \tab DisplayMode \tab TRUE \cr
   escapeDisabled \tab escape-disabled \tab When \code{true}, disables the default close on escape behavior. \tab boolean \tab TRUE \cr
   focusTrapDisabled \tab focus-trap-disabled \tab When \code{true}, prevents focus trapping. \tab boolean \tab TRUE \cr
   focusTrapOptions \tab NA \tab Specifies custom focus trap configuration on the component, where  \code{"allowOutsideClick"} allows outside clicks, \code{"initialFocus"} enables initial focus, \code{"returnFocusOnDeactivate"} returns focus when not active, \code{"extraContainers"} specifies additional focusable elements external to the trap, such as 3rd-party components appending elements to the document body, and \code{"setReturnFocus"} customizes the element to which focus is returned when the trap is deactivated. Return \code{false} to prevent focus return, or \code{undefined} to use the default behavior (returning focus to the element focused before activation). \tab Check API reference \tab FALSE \cr
   height \tab height \tab Specifies the component's height. \tab Height \tab TRUE \cr
   heightScale \tab height-scale \tab When \code{position} is \code{"block-start"} or \code{"block-end"}, specifies the component's height. \tab Scale \tab TRUE \cr
   label \tab label \tab Specifies an accessible label for the component. \tab string \tab FALSE \cr
   messageOverrides \tab NA \tab Overrides individual strings used by the component. \tab Check API reference \tab FALSE \cr
   open \tab open \tab When \code{true}, displays and positions the component. \tab boolean \tab TRUE \cr
   outsideCloseDisabled \tab outside-close-disabled \tab When \code{true}, disables closing the component when the area outside of the component is clicked. \tab boolean \tab TRUE \cr
   position \tab position \tab Determines where the component will be positioned. \tab LogicalFlowPosition \tab TRUE \cr
   resizable \tab resizable \tab When \code{true}, the component is resizable. \tab boolean \tab TRUE \cr
   topLayerDisabled \tab top-layer-disabled \tab When \code{true} and the component is \code{open}, disables top layer placement.  Only set this if you need complex z-index control or if top layer placement causes conflicts with third-party components. \tab boolean \tab TRUE \cr
   width \tab width \tab Specifies the component's width. \tab Check API reference \tab TRUE \cr
   widthScale \tab width-scale \tab When \code{position} is \code{"inline-start"} or \code{"inline-end"}, specifies the component's width. \tab Scale \tab TRUE \cr
}

}

\subsection{Events}{

The following events are observed by shiny:\tabular{ll}{
   Event \tab Description \cr
   calciteSheetBeforeClose \tab Fires when the component is requested to be closed and before the closing transition begins. \cr
   calciteSheetBeforeOpen \tab Fires when the component is added to the DOM but not rendered, and before the opening transition begins. \cr
   calciteSheetClose \tab Fires when the component is closed and animation is complete. \cr
   calciteSheetOpen \tab Fires when the component is open and animation is complete. \cr
}

}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding custom content. \cr
}

}
}
\examples{
calcite_sheet()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/sheet/}{Official Documentation}
}
