% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_shell_center_row}
\alias{calcite_shell_center_row}
\title{Calcite Shell Center Row}
\usage{
calcite_shell_center_row(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-shell-center-row>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/shell-center-row/}{Official Documentation}
}
