% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_stepper_item}
\alias{calcite_stepper_item}
\title{Create a StepperItem component}
\usage{
calcite_stepper_item(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a StepperItem component
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   complete \tab complete \tab When \code{true}, completes the step. \tab boolean \tab TRUE \cr
   description \tab description \tab Specifies a description for the component. Displays below the header text. \tab string \tab FALSE \cr
   disabled \tab disabled \tab When \code{true}, prevents interaction and decreases the component's opacity. \tab boolean \tab TRUE \cr
   error \tab error \tab When \code{true}, the component contains an error that requires resolution from the user. \tab boolean \tab TRUE \cr
   heading \tab heading \tab Specifies the component's heading text. \tab string \tab FALSE \cr
   iconFlipRtl \tab icon-flip-rtl \tab When \code{true}, the icon will be flipped when the element direction is right-to-left (\code{"rtl"}). \tab boolean \tab TRUE \cr
   messageOverrides \tab NA \tab Overrides individual strings used by the component. \tab Check API reference \tab FALSE \cr
   selected \tab selected \tab When \code{true}, the component is selected. \tab boolean \tab TRUE \cr
}

}

\subsection{Events}{

The following events are observed by shiny:\tabular{ll}{
   Event \tab Description \cr
   calciteStepperItemSelect \tab Fires when the active \code{calcite-stepper-item} changes. \cr
}

}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding custom content. \cr
}

}
}
\examples{
calcite_stepper_item()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/stepper-item/}{Official Documentation}
}
