% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_swatch}
\alias{calcite_swatch}
\title{Create a Swatch component}
\usage{
calcite_swatch(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a Swatch component
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   color \tab color \tab Specifies the component's color \tab string \tab FALSE \cr
   disabled \tab disabled \tab When \code{true}, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab TRUE \cr
   label \tab label \tab Accessible name for the component. \tab string \tab FALSE \cr
   scale \tab scale \tab Specifies the size of the component. When contained in a parent \code{calcite-swatch-group} inherits the parent's \code{scale} value. \tab "l" | "m" | "s" \tab TRUE \cr
   selected \tab selected \tab When \code{true}, the component is selected. \tab boolean \tab TRUE \cr
   value \tab value \tab The component's value. \tab any \tab FALSE \cr
}

}

\subsection{Events}{

The following events are observed by shiny:\tabular{ll}{
   Event \tab Description \cr
   calciteSwatchSelect \tab Fires when the selected state of the component changes. \cr
}

}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   image \tab A slot for adding an image or pattern. \cr
}

}
}
\examples{
calcite_swatch()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/swatch/}{Official Documentation}
}
