% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tab}
\alias{calcite_tab}
\title{Calcite Tab}
\usage{
calcite_tab(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tab>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   selected \tab selected \tab When true, the component’s contents are selected. Only one tab can be selected within the calcite-tabs parent. \tab boolean \tab false \cr
   tab \tab tab \tab Specifies a unique name for the component. When specified, use the same value on the calcite-tab-title. \tab string \tab undefined \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding custom content. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tab/}{Official Documentation}
}
