% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tab_nav}
\alias{calcite_tab_nav}
\title{Calcite Tab Nav}
\usage{
calcite_tab_nav(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tab-nav>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tab-nav/}{Official Documentation}
}
