% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R, R/table.R
\name{calcite_table_header}
\alias{calcite_table_header}
\title{Create a TableHeader component}
\usage{
calcite_table_header(heading, description = NULL, alignment = "start")

calcite_table_header(heading, description = NULL, alignment = "start")
}
\arguments{
\item{heading}{Header text to display}

\item{description}{Optional description text below heading}

\item{alignment}{Alignment: "start", "center", or "end" (default: "start")}

\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}

A calcite-table-header tag
}
\description{
Creates a header cell for use in calcite_table headers.
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   alignment \tab alignment \tab Specifies the alignment of the component. \tab Alignment \tab TRUE \cr
   colSpan \tab col-span \tab Specifies the number of columns the component should span. \tab number \tab TRUE \cr
   description \tab description \tab Specifies a description for the component. Displays below the \code{heading}. \tab string \tab TRUE \cr
   heading \tab heading \tab Specifies the component's heading text. Displays above the \code{description}. \tab string \tab TRUE \cr
   messageOverrides \tab NA \tab Overrides individual strings used by the component. \tab Check API reference \tab FALSE \cr
   rowSpan \tab row-span \tab Specifies the number of rows the component should span. \tab number \tab TRUE \cr
}

}
}
\examples{
calcite_table_header()
calcite_table_header(
  heading = "Name",
  description = "Person's full name"
)
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/table-header/}{Official Documentation}
}
