% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tip_group}
\alias{calcite_tip_group}
\title{Calcite Tip Group}
\usage{
calcite_tip_group(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tip-group>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   groupTitle \tab group-title \tab The component header text for all nested calcite-tips. \tab string \tab undefined \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-tips. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tip-group/}{Official Documentation}
}
