% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcite_version}
\alias{calcite_version}
\title{Calcite components version}
\usage{
calcite_version()
}
\value{
a character scalar
}
\description{
Reports the currently used calcite component version.
}
\examples{
calcite_version()
}
