% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcite.R
\name{calciteInput}
\alias{calciteInput}
\title{calcite}
\usage{
calciteInput(..., .noWS = NULL, .renderHook = NULL)
}
\arguments{
\item{inputId}{Id of input.}

\item{value}{Initial value.}
}
\description{
Incremental button.
}
\examples{
library(shiny)

ui <- fluidPage(
  calciteInput("theId", 0)
)

server <- function(input, output) {
  observeEvent(input$theId, {
    print(input$theId)
  })
}

if (interactive()) {
  shinyApp(ui, server)
}

}
