% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{examples}
\alias{examples}
\alias{list_examples}
\alias{run_example}
\alias{open_example}
\title{Calcite examples}
\usage{
list_examples()

run_example(name)

open_example(name = NULL)
}
\arguments{
\item{name}{the example name (without \code{.R} extension). See \code{\link[=list_examples]{list_examples()}}.}
}
\description{
Calcite examples
}
\examples{
list_examples()

if (interactive()) {
  open_example()
  run_example("calcite-checkbox")
}
}
