% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_accordion_item}
\alias{calcite_accordion_item}
\title{Calcite Accordion Item}
\usage{
calcite_accordion_item(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-accordion-item>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   description \tab description \tab Specifies a description for the component. \tab string \tab undefined \cr
   expanded \tab expanded \tab When true, the component is expanded. \tab boolean \tab false \cr
   heading \tab heading \tab Specifies heading text for the component. \tab string \tab undefined \cr
   iconEnd \tab icon-end \tab Specifies an icon to display at the end of the component. \tab string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab Displays the iconStart and/or iconEnd as flipped when the element direction is right-to-left ("rtl"). \tab "both" | "end" | "start" \tab undefined \cr
   iconStart \tab icon-start \tab Specifies an icon to display at the start of the component. \tab string \tab undefined \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding custom content, including nested calcite-accordion-items. \cr
   "actions-end" \tab A slot for adding calcite-actions or content to the end side of the component’s header. \cr
   "actions-start" \tab A slot for adding calcite-actions or content to the start side of the component’s header. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/accordion-item/}{Official Documentation}
}
