% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_action_menu}
\alias{calcite_action_menu}
\title{Calcite Action Menu}
\usage{
calcite_action_menu(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-action-menu>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   appearance \tab appearance \tab Specifies the appearance of the component. \tab "solid" | "transparent" \tab "solid" \cr
   expanded \tab expanded \tab When true, the component is expanded. \tab boolean \tab false \cr
   flipPlacements \tab – \tab Specifies the component’s fallback slotted content placement when it’s initial or specified placement has insufficient space available. \tab FlipPlacement[] \tab undefined \cr
   label (required) \tab label \tab Specifies the text string for the component. \tab string \tab undefined \cr
   open \tab open \tab When true, the component is open. \tab boolean \tab false \cr
   overlayPositioning \tab overlay-positioning \tab Determines the type of positioning to use for the overlaid content. Using "absolute" will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container’s layout. "fixed" should be used to escape an overflowing parent container, or when the reference element’s position CSS property is "fixed". \tab "absolute" | "fixed" \tab "absolute" \cr
   placement \tab placement \tab Determines where the component will be positioned relative to the referenceElement. \tab "auto" | "top" | "bottom" | "right" | "left" | "top-start" | "top-end" | "bottom-start" | "bottom-end" | "right-start" | "right-end" | "left-start" | "left-end" | "leading" | "trailing" | "auto-start" | "auto-end" | "leading-start" | "leading-end" | "trailing-end" | "trailing-start" \tab "auto" \cr
   scale \tab scale \tab Specifies the size of the component’s trigger calcite-action. \tab "l" | "m" | "s" \tab undefined \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteActionMenuOpen \tab Fires when the open property is toggled. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding calcite-actions. \cr
   "tooltip" \tab A slot for adding an tooltip for the menu. \cr
   "trigger" \tab A slot for adding a calcite-action to trigger opening the menu. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/action-menu/}{Official Documentation}
}
