% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_block_group}
\alias{calcite_block_group}
\title{Create a BlockGroup component}
\usage{
calcite_block_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a BlockGroup component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   canPull \tab NA \tab When provided, the method will be called to determine whether the element is sortable in the component. \tab (detail: BlockDragDetail) => boolean \tab FALSE \cr
   canPut \tab NA \tab When provided, the method will be called to determine whether the element can be added from another component. \tab (detail: BlockDragDetail) => boolean \tab FALSE \cr
   disabled \tab disabled \tab When \code{true}, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab TRUE \cr
   dragEnabled \tab drag-enabled \tab When \code{true}, \code{calcite-block}s are sortable via a draggable button. \tab boolean \tab TRUE \cr
   group \tab group \tab The block-group's group identifier.  To drag elements from one group into another, both groups must have the same group value. \tab string \tab TRUE \cr
   label \tab label \tab Specifies an accessible name for the component.  When \code{dragEnabled} is \code{true} and multiple group sorting is enabled with \code{group}, specifies the component's name for dragging between groups. \tab string \tab FALSE \cr
   loading \tab loading \tab When \code{true}, a busy indicator is displayed. \tab boolean \tab TRUE \cr
}

\subsection{Events}{

The following events are observed by shiny:\tabular{ll}{
   Event \tab Description \cr
   calciteBlockGroupDragEnd \tab Fires when the component's dragging has ended. \cr
   calciteBlockGroupDragStart \tab Fires when the component's dragging has started. \cr
   calciteBlockGroupOrderChange \tab Fires when the component's item order changes. \cr
}

}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding \code{calcite-block} elements. \cr
}

}
}
\examples{
calcite_block_group()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/block-group/}{Official Documentation}
}
