% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_card_group}
\alias{calcite_card_group}
\title{Calcite Card Group}
\usage{
calcite_card_group(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-card-group>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   selectedItems \tab – \tab Specifies the component’s selected items. \tab HTMLCalciteCardElement[] \tab [] \cr
   selectionMode \tab selection-mode \tab Specifies the selection mode of the component. \tab "multiple" | "none" | "single" | "single-persist" \tab "none" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteCardGroupSelect \tab Emits when the component’s selection changes and the selectionMode is not none. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding one or more calcite-cards. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/card-group/}{Official Documentation}
}
