% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_carousel_item}
\alias{calcite_carousel_item}
\title{Calcite Carousel Item}
\usage{
calcite_carousel_item(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-carousel-item>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   selected \tab selected \tab When true, the component is selected. \tab boolean \tab false \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding content. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/carousel-item/}{Official Documentation}
}
