% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_checkbox}
\alias{calcite_checkbox}
\title{Calcite Checkbox}
\usage{
calcite_checkbox(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-checkbox>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/checkbox/}{Official Documentation}
}
