% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_combobox}
\alias{calcite_combobox}
\title{Calcite Combobox}
\usage{
calcite_combobox(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-combobox>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/combobox/}{Official Documentation}
}
