% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_combobox_item}
\alias{calcite_combobox_item}
\title{Calcite Combobox Item}
\usage{
calcite_combobox_item(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-combobox-item>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   active \tab active \tab When true, the component is active. \tab boolean \tab false \cr
   ancestors \tab – \tab Specifies the parent and grandparent items, which are set on calcite-combobox. \tab ComboboxChildElement[] \tab undefined \cr
   description \tab description \tab A description for the component, which displays below the label. \tab string \tab undefined \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   filterDisabled \tab filter-disabled \tab When true, omits the component from the calcite-combobox filtered search results. \tab boolean \tab undefined \cr
   guid \tab guid \tab The id attribute of the component. When omitted, a globally unique identifier is used. \tab string \tab guid() \cr
   heading \tab heading \tab The component’s text. \tab string \tab undefined \cr
   icon \tab icon \tab Specifies an icon to display. \tab string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab When true, the icon will be flipped when the element direction is right-to-left ("rtl"). \tab boolean \tab false \cr
   label \tab label \tab The component’s label. \tab any \tab undefined \cr
   metadata \tab – \tab Provides additional metadata to the component used in filtering. \tab { \code{x: string}: unknown; } \tab undefined \cr
   selected \tab selected \tab When true, the component is selected. \tab boolean \tab false \cr
   shortHeading \tab short-heading \tab The component’s short heading. When provided, the short heading will be displayed in the component’s selection. It is recommended to use 5 characters or fewer. \tab string \tab undefined \cr
   textLabel (required) \tab text-label \tab DEPRECATED Use heading instead.The component’s text. \tab string \tab undefined \cr
   value (required) \tab value \tab The component’s value. \tab any \tab undefined \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteComboboxItemChange \tab Fires whenever the component is selected or unselected. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding nested calcite-combobox-items. \cr
   "content-end" \tab A slot for adding non-actionable elements after the component’s content. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/combobox-item/}{Official Documentation}
}
