% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_combobox_item_group}
\alias{calcite_combobox_item_group}
\title{Calcite Combobox Item Group}
\usage{
calcite_combobox_item_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-combobox-item-group>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   ancestors \tab – \tab Specifies the parent and grandparent calcite-combobox-items, which are set on calcite-combobox. \tab ComboboxChildElement[] \tab undefined \cr
   label (required) \tab label \tab Specifies the title of the component. \tab string \tab undefined \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-combobox-items. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/combobox-item-group/}{Official Documentation}
}
