% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_date_picker_day}
\alias{calcite_date_picker_day}
\title{Calcite Date Picker Day}
\usage{
calcite_date_picker_day(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-date-picker-day>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   active \tab active \tab When true, the component is active. \tab boolean \tab false \cr
   currentMonth \tab current-month \tab Date is in the current month. \tab boolean \tab false \cr
   day (required) \tab day \tab Day of the month to be shown. \tab number \tab undefined \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   endOfRange \tab end-of-range \tab Date is the end of date range. \tab boolean \tab false \cr
   highlighted \tab highlighted \tab Date is currently highlighted as part of the range, \tab boolean \tab false \cr
   range \tab range \tab When true, activates the component’s range mode to allow a start and end date. \tab boolean \tab false \cr
   rangeHover \tab range-hover \tab Date is being hovered and within the set range. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab undefined \cr
   selected \tab selected \tab When true, the component is selected. \tab boolean \tab false \cr
   startOfRange \tab start-of-range \tab Date is the start of date range. \tab boolean \tab false \cr
   value \tab – \tab The component’s value. \tab Date \tab undefined \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteInternalDaySelect \tab Fires when user selects day. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/date-picker-day/}{Official Documentation}
}
