% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_date_picker_month}
\alias{calcite_date_picker_month}
\title{Calcite Date Picker Month}
\usage{
calcite_date_picker_month(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-date-picker-month>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/date-picker-month/}{Official Documentation}
}
