% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_date_picker_month}
\alias{calcite_date_picker_month}
\title{Create a DatePickerMonth component}
\usage{
calcite_date_picker_month(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a DatePickerMonth component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   activeDate \tab NA \tab The currently active Date. \tab Date \tab FALSE \cr
   endDate \tab NA \tab End date currently active. \tab Date \tab FALSE \cr
   headingLevel \tab heading-level \tab Specifies the number at which section headings should start. \tab 1 | 2 | 3 | 4 | 5 | 6 \tab TRUE \cr
   hoverRange \tab NA \tab The range of dates currently being hovered. \tab HoverRange \tab FALSE \cr
   max \tab NA \tab Specifies the latest allowed date (\code{"yyyy-mm-dd"}). \tab Date \tab FALSE \cr
   min \tab NA \tab Specifies the earliest allowed date (\code{"yyyy-mm-dd"}). \tab Date \tab FALSE \cr
   monthStyle \tab month-style \tab Specifies the monthStyle used by the component. \tab "abbreviated" | "wide" \tab FALSE \cr
   range \tab range \tab When \code{true}, activates the component's range mode which renders two calendars for selecting ranges of dates. \tab boolean \tab TRUE \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab TRUE \cr
   selectedDate \tab NA \tab Already selected date. \tab Date \tab FALSE \cr
   startDate \tab NA \tab Start date currently active. \tab Date \tab FALSE \cr
}
}
\examples{
calcite_date_picker_month()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/date-picker-month/}{Official Documentation}
}
