% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_dialog}
\alias{calcite_dialog}
\title{Calcite Dialog}
\usage{
calcite_dialog(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-dialog>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   beforeClose \tab – \tab Passes a function to run before the component closes. \tab () => Promise\if{html}{\out{<void>}} \tab undefined \cr
   closeDisabled \tab close-disabled \tab When true, disables the component’s close button. \tab boolean \tab false \cr
   description \tab description \tab A description for the component. \tab string \tab undefined \cr
   dragEnabled \tab drag-enabled \tab When true, the component is draggable. \tab boolean \tab false \cr
   escapeDisabled \tab escape-disabled \tab When true, disables the default close on escape behavior. By default, an open dialog can be dismissed by pressing the Esc key. \tab boolean \tab false \cr
   heading \tab heading \tab The component header text. \tab string \tab undefined \cr
   headingLevel \tab heading-level \tab Specifies the heading level of the component’s heading for proper document structure, without affecting visual styling. \tab 1 | 2 | 3 | 4 | 5 | 6 \tab undefined \cr
   kind \tab kind \tab Specifies the kind of the component, which will style the top border. \tab "brand" | "danger" | "info" | "success" | "warning" \tab undefined \cr
   loading \tab loading \tab When true, a busy indicator is displayed. \tab boolean \tab false \cr
   menuOpen \tab menu-open \tab When true, the action menu items in the header-menu-actions slot are open. \tab boolean \tab false \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab DialogMessages \tab undefined \cr
   modal \tab modal \tab When true, displays a scrim blocking interaction underneath the component. \tab boolean \tab false \cr
   open \tab open \tab When true, displays and positions the component. \tab boolean \tab false \cr
   outsideCloseDisabled \tab outside-close-disabled \tab When true, disables the closing of the component when clicked outside. \tab boolean \tab false \cr
   overlayPositioning \tab overlay-positioning \tab Determines the type of positioning to use for the overlaid content. Using "absolute" will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container’s layout. "fixed" should be used to escape an overflowing parent container, or when the reference element’s position CSS property is "fixed". \tab "absolute" | "fixed" \tab "absolute" \cr
   placement \tab placement \tab Specifies the placement of the dialog. \tab "bottom" | "bottom-end" | "bottom-start" | "center" | "cover" | "top" | "top-end" | "top-start" \tab "center" \cr
   resizable \tab resizable \tab When true, the component is resizable. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   widthScale \tab width-scale \tab Specifies the width of the component. \tab "l" | "m" | "s" \tab "m" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteDialogBeforeClose \tab Fires when the component is requested to be closed and before the closing transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteDialogBeforeOpen \tab Fires when the component is added to the DOM but not rendered, and before the opening transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteDialogClose \tab Fires when the component is closed and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteDialogOpen \tab Fires when the component is open and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteDialogScroll \tab Fires when the content is scrolled. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding content. \cr
   "action-bar" \tab A slot for adding a calcite-action-bar to the component. \cr
   "alerts" \tab A slot for adding calcite-alerts to the component. \cr
   "content" \tab A slot for adding custom content. \cr
   "content-bottom" \tab A slot for adding content below the unnamed (default) slot and - if populated - the footer slot. \cr
   "content-top" \tab A slot for adding content above the unnamed (default) slot and - if populated - below the action-bar slot. \cr
   "fab" \tab A slot for adding a calcite-fab (floating action button) to perform an action. \cr
   "footer" \tab A slot for adding custom content to the component’s footer. Should not be used with the "footer-start" or "footer-end" slots. \cr
   "footer-end" \tab A slot for adding a trailing footer custom content. Should not be used with the "footer" slot. \cr
   "footer-start" \tab A slot for adding a leading footer custom content. Should not be used with the "footer" slot. \cr
   "header-actions-end" \tab A slot for adding actions or content to the ending side of the component’s header. \cr
   "header-actions-start" \tab A slot for adding actions or content to the starting side of the component’s header. \cr
   "header-content" \tab A slot for adding custom content to the component’s header. \cr
   "header-menu-actions" \tab A slot for adding an overflow menu with actions inside a calcite-dropdown. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/dialog/}{Official Documentation}
}
