% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_dropdown_item}
\alias{calcite_dropdown_item}
\title{Calcite Dropdown Item}
\usage{
calcite_dropdown_item(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-dropdown-item>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   href \tab href \tab Specifies the URL of the linked resource, which can be set as an absolute or relative path. Determines if the component will render as an anchor. \tab string \tab undefined \cr
   iconEnd \tab icon-end \tab Specifies an icon to display at the end of the component. \tab string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab Displays the iconStart and/or iconEnd as flipped when the element direction is right-to-left ("rtl"). \tab "both" | "end" | "start" \tab undefined \cr
   iconStart \tab icon-start \tab Specifies an icon to display at the start of the component. \tab string \tab undefined \cr
   label \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   rel \tab rel \tab Specifies the relationship to the linked document defined in href. \tab string \tab undefined \cr
   selected \tab selected \tab When true, the component is selected. \tab boolean \tab false \cr
   target \tab target \tab Specifies the frame or window to open the linked document. \tab string \tab undefined \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteDropdownItemSelect \tab Fires when the component is selected. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding text. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/dropdown-item/}{Official Documentation}
}
