% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_filter}
\alias{calcite_filter}
\title{Calcite Filter}
\usage{
calcite_filter(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-filter>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   filterProps \tab – \tab Specifies the properties to match against when filtering. This will only apply when value is an object. If not set, all properties will be matched. \tab string[] \tab undefined \cr
   filteredItems \tab – \tab The component’s resulting items after filtering. \tab object[] \tab [] \cr
   items \tab – \tab Defines the items to filter. The component uses the values as the starting point, and returns items that contain the string entered in the input, using a partial match and recursive search. This property is needed to conduct filtering. \tab object[] \tab [] \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab FilterMessages \tab undefined \cr
   placeholder \tab placeholder \tab Specifies placeholder text for the input element. \tab string \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   value \tab value \tab The component’s value. \tab string \tab "" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteFilterChange \tab Fires when the filter text changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/filter/}{Official Documentation}
}
