% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_flow}
\alias{calcite_flow}
\title{Calcite Flow}
\usage{
calcite_flow(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-flow>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-flow-item elements to the component. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/flow/}{Official Documentation}
}
