% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_input_message}
\alias{calcite_input_message}
\title{Calcite Input Message}
\usage{
calcite_input_message(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-input-message>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/input-message/}{Official Documentation}
}
