% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_input_text}
\alias{calcite_input_text}
\title{Calcite Input Text}
\usage{
calcite_input_text(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-input-text>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/input-text/}{Official Documentation}
}
