% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_input_text}
\alias{calcite_input_text}
\title{Calcite Input Text}
\usage{
calcite_input_text(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-input-text>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   alignment \tab alignment \tab Specifies the text alignment of the component’s value. \tab "end" | "start" \tab "start" \cr
   autocomplete \tab autocomplete \tab Specifies the type of content to autocomplete, for use in forms. Read the native attribute’s documentation on MDN for more info. \tab string \tab undefined \cr
   autofocus \tab autofocus \tab Adds global prop, missing from Stencil’s HTMLElement type, see https://github.com/ionic-team/stencil/issues/5726 \tab boolean \tab undefined \cr
   clearable \tab clearable \tab When true, a clear button is displayed when the component has a value. \tab boolean \tab false \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   form \tab form \tab The id of the form that will be associated with the component. When not set, the component will be associated with its ancestor form element, if any. \tab string \tab undefined \cr
   icon \tab icon \tab Specifies an icon to display. \tab boolean | string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab When true, the icon will be flipped when the element direction is right-to-left ("rtl"). \tab boolean \tab false \cr
   label \tab label \tab Accessible name for the component’s button or hyperlink. \tab string \tab undefined \cr
   loading \tab loading \tab When true, the component is in the loading state and calcite-progress is displayed. \tab boolean \tab false \cr
   maxLength \tab max-length \tab Specifies the maximum length of text for the component’s value. \tab number \tab undefined \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab InputTextMessages \tab undefined \cr
   minLength \tab min-length \tab Specifies the minimum length of text for the component’s value. \tab number \tab undefined \cr
   name \tab name \tab Specifies the name of the component. Required to pass the component’s value on form submission. \tab string \tab undefined \cr
   pattern \tab pattern \tab Specifies a regex pattern the component’s value must match for validation. Read the native attribute’s documentation on MDN for more info. \tab string \tab undefined \cr
   placeholder \tab placeholder \tab Specifies placeholder text for the component. \tab string \tab undefined \cr
   prefixText \tab prefix-text \tab Adds text to the start of the component. \tab string \tab undefined \cr
   readOnly \tab read-only \tab When true, the component’s value can be read, but cannot be modified. \tab boolean \tab false \cr
   required \tab required \tab When true, the component must have a value in order for the form to submit. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   status \tab status \tab Specifies the status of the input field, which determines message and icons. \tab "idle" | "invalid" | "valid" \tab "idle" \cr
   suffixText \tab suffix-text \tab Adds text to the end of the component. \tab string \tab undefined \cr
   validationIcon \tab validation-icon \tab Specifies the validation icon to display under the component. \tab boolean | string \tab undefined \cr
   validationMessage \tab validation-message \tab Specifies the validation message to display under the component. \tab string \tab undefined \cr
   validity \tab – \tab The current validation state of the component. \tab { valid: boolean; badInput: boolean; customError: boolean; patternMismatch: boolean; rangeOverflow: boolean; rangeUnderflow: boolean; stepMismatch: boolean; tooLong: boolean; tooShort: boolean; typeMismatch: boolean; valueMissing: boolean; } \tab {     valid: false,     badInput: false,     customError: false,     patternMismatch: false,     rangeOverflow: false,     rangeUnderflow: false,     stepMismatch: false,     tooLong: false,     tooShort: false,     typeMismatch: false,     valueMissing: false,   } \cr
   value \tab value \tab The component’s value. \tab string \tab "" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteInputTextChange \tab Fires each time a new value is typed and committed. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteInputTextInput \tab Fires each time a new value is typed. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   "action" \tab A slot for positioning a button next to the component. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/input-text/}{Official Documentation}
}
