% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_input_time_zone}
\alias{calcite_input_time_zone}
\title{Calcite Input Time Zone}
\usage{
calcite_input_time_zone(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-input-time-zone>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   clearable \tab clearable \tab When true, an empty value (null) will be allowed as a value. When false, an offset or name value is enforced, and clearing the input or blurring will restore the last valid value. \tab boolean \tab false \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   form \tab form \tab The id of the form that will be associated with the component. When not set, the component will be associated with its ancestor form element, if any. \tab string \tab undefined \cr
   maxItems \tab max-items \tab Specifies the component’s maximum number of options to display before displaying a scrollbar. \tab number \tab 0 \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab InputTimeZoneMessages \tab undefined \cr
   mode \tab mode \tab This specifies the type of value and the associated options presented to the user: Using "offset" will provide options that show timezone offsets. Using "name" will provide options that show the IANA time zone names. \tab "name" | "offset" | "region" \tab "offset" \cr
   name \tab name \tab Specifies the name of the component. Required to pass the component’s value on form submission. \tab string \tab undefined \cr
   offsetStyle \tab offset-style \tab Specifies how the offset will be displayed, where "user" uses UTC or GMT depending on the user’s locale, "gmt" always uses GMT, and "utc" always uses UTC. This only applies to the offset mode. \tab "gmt" | "user" | "utc" \tab "user" \cr
   open \tab open \tab When true, displays and positions the component. \tab boolean \tab false \cr
   overlayPositioning \tab overlay-positioning \tab Determines the type of positioning to use for the overlaid content. Using "absolute" will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container’s layout. "fixed" should be used to escape an overflowing parent container, or when the reference element’s position CSS property is "fixed". \tab "absolute" | "fixed" \tab "absolute" \cr
   readOnly \tab read-only \tab When true, the component’s value can be read, but controls are not accessible and the value cannot be modified. \tab boolean \tab false \cr
   referenceDate \tab reference-date \tab This date will be used as a reference to Daylight Savings Time when creating time zone item groups. It can be either a Date instance or a string in ISO format ("YYYY-MM-DD", "YYYY-MM-DDTHH:MM:SS.SSSZ"). \tab Date | string \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   status \tab status \tab Specifies the status of the input field, which determines message and icons. \tab "idle" | "invalid" | "valid" \tab "idle" \cr
   validationIcon \tab validation-icon \tab Specifies the validation icon to display under the component. \tab boolean | string \tab undefined \cr
   validationMessage \tab validation-message \tab Specifies the validation message to display under the component. \tab string \tab undefined \cr
   validity \tab – \tab The current validation state of the component. \tab { valid: boolean; badInput: boolean; customError: boolean; patternMismatch: boolean; rangeOverflow: boolean; rangeUnderflow: boolean; stepMismatch: boolean; tooLong: boolean; tooShort: boolean; typeMismatch: boolean; valueMissing: boolean; } \tab {     valid: false,     badInput: false,     customError: false,     patternMismatch: false,     rangeOverflow: false,     rangeUnderflow: false,     stepMismatch: false,     tooLong: false,     tooShort: false,     typeMismatch: false,     valueMissing: false,   } \cr
   value \tab value \tab The component’s value, where the value is the time zone offset or the difference, in minutes, between the selected time zone and UTC. If no value is provided, the user’s time zone offset will be selected by default. \tab string \tab undefined \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteInputTimeZoneBeforeClose \tab Fires when the component is requested to be closed and before the closing transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteInputTimeZoneBeforeOpen \tab Fires when the component is added to the DOM but not rendered, and before the opening transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteInputTimeZoneChange \tab Fires when the component’s value changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteInputTimeZoneClose \tab Fires after the component is closed and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteInputTimeZoneOpen \tab Fires after the component is opened and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/input-time-zone/}{Official Documentation}
}
