% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_label}
\alias{calcite_label}
\title{Calcite Label}
\usage{
calcite_label(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-label>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/label/}{Official Documentation}
}
