% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_label}
\alias{calcite_label}
\title{Calcite Label}
\usage{
calcite_label(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-label>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   alignment \tab alignment \tab Specifies the text alignment of the component. \tab "center" | "end" | "start" \tab "start" \cr
   for \tab for \tab Specifies the id of the component the label is bound to. Use when the component the label is bound to does not reside within the component. \tab string \tab undefined \cr
   layout \tab layout \tab Defines the layout of the label in relation to the component. Use "inline" positions to wrap the label and component on the same line. \tab "default" | "inline" | "inline-space-between" \tab "default" \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding text and a component that can be labeled. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/label/}{Official Documentation}
}
