% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_list_item}
\alias{calcite_list_item}
\title{Calcite List Item}
\usage{
calcite_list_item(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-list-item>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/list-item/}{Official Documentation}
}
