% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_list_item_group}
\alias{calcite_list_item_group}
\title{Create a ListItemGroup component}
\usage{
calcite_list_item_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a ListItemGroup component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   disabled \tab disabled \tab When \code{true}, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab TRUE \cr
   heading \tab heading \tab The header text for all nested \code{calcite-list-item} rows. \tab string \tab TRUE \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab TRUE \cr
}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding \code{calcite-list-item} and \code{calcite-list-item-group} elements. \cr
}

}
}
\examples{
calcite_list_item_group()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/list-item-group/}{Official Documentation}
}
