% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_list_item_group}
\alias{calcite_list_item_group}
\title{Calcite List Item Group}
\usage{
calcite_list_item_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-list-item-group>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/list-item-group/}{Official Documentation}
}
