% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_menu}
\alias{calcite_menu}
\title{Create a Menu component}
\usage{
calcite_menu(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a Menu component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   label \tab label \tab Accessible name for the component. \tab string \tab FALSE \cr
   layout \tab layout \tab Specifies the layout of the component. \tab "horizontal" | "vertical" \tab TRUE \cr
   messageOverrides \tab NA \tab Use this property to override individual strings used by the component. \tab Check API reference \tab FALSE \cr
}
}
\examples{
calcite_menu()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/menu/}{Official Documentation}
}
