% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_modal}
\alias{calcite_modal}
\title{Calcite Modal}
\usage{
calcite_modal(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-modal>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/modal/}{Official Documentation}
}
