% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_navigation_logo}
\alias{calcite_navigation_logo}
\title{Calcite Navigation Logo}
\usage{
calcite_navigation_logo(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-navigation-logo>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   active \tab active \tab When true, the component is highlighted. \tab boolean \tab undefined \cr
   description \tab description \tab A description for the component, which displays below the heading. \tab string \tab undefined \cr
   heading \tab heading \tab Specifies heading text for the component, such as a product or organization name. \tab string \tab undefined \cr
   headingLevel \tab heading-level \tab Specifies the heading level of the component’s heading for proper document structure, without affecting visual styling. \tab 1 | 2 | 3 | 4 | 5 | 6 \tab undefined \cr
   href \tab href \tab Specifies the URL destination of the component, which can be set as an absolute or relative path. \tab string \tab undefined \cr
   icon \tab icon \tab Specifies an icon to display. \tab string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab When true, the icon will be flipped when the element direction is right-to-left ("rtl"). \tab boolean \tab false \cr
   label \tab label \tab Describes the appearance or function of the thumbnail. If no label is provided, context will not be provided to assistive technologies. \tab string \tab undefined \cr
   rel \tab rel \tab Defines the relationship between the href value and the current document. \tab string \tab undefined \cr
   target \tab target \tab Specifies where to open the linked document defined in the href property. \tab string \tab undefined \cr
   thumbnail \tab thumbnail \tab Specifies the src to an image. \tab string \tab undefined \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/navigation-logo/}{Official Documentation}
}
