% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_navigation_user}
\alias{calcite_navigation_user}
\title{Calcite Navigation User}
\usage{
calcite_navigation_user(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-navigation-user>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/navigation-user/}{Official Documentation}
}
