% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_panel}
\alias{calcite_panel}
\title{Calcite Panel}
\usage{
calcite_panel(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-panel>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/panel/}{Official Documentation}
}
