% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_popover}
\alias{calcite_popover}
\title{Calcite Popover}
\usage{
calcite_popover(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-popover>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   autoClose \tab auto-close \tab When true, clicking outside of the component automatically closes open calcite-popovers. \tab boolean \tab false \cr
   closable \tab closable \tab When true, displays a close button within the component. \tab boolean \tab false \cr
   flipDisabled \tab flip-disabled \tab When true, prevents flipping the component’s placement when overlapping its referenceElement. \tab boolean \tab false \cr
   flipPlacements \tab – \tab Specifies the component’s fallback placement when it’s initial or specified placement has insufficient space available. \tab FlipPlacement[] \tab undefined \cr
   focusTrapDisabled \tab focus-trap-disabled \tab When true, prevents focus trapping. \tab boolean \tab false \cr
   heading \tab heading \tab The component header text. \tab string \tab undefined \cr
   headingLevel \tab heading-level \tab Specifies the heading level of the component’s heading for proper document structure, without affecting visual styling. \tab 1 | 2 | 3 | 4 | 5 | 6 \tab undefined \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab PopoverMessages \tab undefined \cr
   offsetDistance \tab offset-distance \tab Offsets the position of the popover away from the referenceElement. \tab number \tab defaultOffsetDistance \cr
   offsetSkidding \tab offset-skidding \tab Offsets the position of the component along the referenceElement. \tab number \tab 0 \cr
   open \tab open \tab When true, displays and positions the component. \tab boolean \tab false \cr
   overlayPositioning \tab overlay-positioning \tab Determines the type of positioning to use for the overlaid content. Using "absolute" will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container’s layout. "fixed" value should be used to escape an overflowing parent container, or when the reference element’s position CSS property is "fixed". \tab "absolute" | "fixed" \tab "absolute" \cr
   placement \tab placement \tab Determines where the component will be positioned relative to the referenceElement. \tab "auto" | "top" | "bottom" | "right" | "left" | "top-start" | "top-end" | "bottom-start" | "bottom-end" | "right-start" | "right-end" | "left-start" | "left-end" | "leading" | "trailing" | "auto-start" | "auto-end" | "leading-start" | "leading-end" | "trailing-end" | "trailing-start" \tab defaultPopoverPlacement \cr
   pointerDisabled \tab pointer-disabled \tab When true, removes the caret pointer. \tab boolean \tab false \cr
   referenceElement (required) \tab reference-element \tab The referenceElement used to position the component according to its placement value. Setting to an HTMLElement is preferred so the component does not need to query the DOM. However, a string id of the reference element can also be used. \tab Element | VirtualElement | string \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   triggerDisabled \tab trigger-disabled \tab When true, disables automatically toggling the component when its referenceElement has been triggered. This property can be set to true to manage when the component is open. \tab boolean \tab false \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calcitePopoverBeforeClose \tab Fires when the component is requested to be closed and before the closing transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calcitePopoverBeforeOpen \tab Fires when the component is added to the DOM but not rendered, and before the opening transition begins. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calcitePopoverClose \tab Fires when the component is closed and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calcitePopoverOpen \tab Fires when the component is open and animation is complete. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding custom content. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/popover/}{Official Documentation}
}
