% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_radio_button_group}
\alias{calcite_radio_button_group}
\title{Calcite Radio Button Group}
\usage{
calcite_radio_button_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-radio-button-group>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   layout \tab layout \tab Defines the layout of the component. \tab "grid" | "horizontal" | "vertical" \tab "horizontal" \cr
   name (required) \tab name \tab Specifies the name of the component on form submission. Must be unique to other component instances. \tab string \tab undefined \cr
   required \tab required \tab When true, the component must have a value in order for the form to submit. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   selectedItem \tab – \tab Specifies the component’s selected item. \tab HTMLCalciteRadioButtonElement \tab null \cr
   status \tab status \tab Specifies the status of the validation message. \tab "idle" | "invalid" | "valid" \tab "idle" \cr
   validationIcon \tab validation-icon \tab Specifies the validation icon to display under the component. \tab boolean | string \tab undefined \cr
   validationMessage \tab validation-message \tab Specifies the validation message to display under the component. \tab string \tab undefined \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteRadioButtonGroupChange \tab Fires when the component has changed. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-radio-buttons. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/radio-button-group/}{Official Documentation}
}
