% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_rating}
\alias{calcite_rating}
\title{Calcite Rating}
\usage{
calcite_rating(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-rating>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/rating/}{Official Documentation}
}
