% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_segmented_control}
\alias{calcite_segmented_control}
\title{Calcite Segmented Control}
\usage{
calcite_segmented_control(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-segmented-control>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/segmented-control/}{Official Documentation}
}
