% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_segmented_control_item}
\alias{calcite_segmented_control_item}
\title{Create a SegmentedControlItem component}
\usage{
calcite_segmented_control_item(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a SegmentedControlItem component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   checked \tab checked \tab When \code{true}, the component is checked. \tab boolean \tab TRUE \cr
   iconEnd \tab icon-end \tab Specifies an icon to display at the end of the component. \tab string \tab TRUE \cr
   iconFlipRtl \tab icon-flip-rtl \tab When \code{true}, the icon will be flipped when the element direction is right-to-left (\code{"rtl"}). \tab boolean \tab TRUE \cr
   iconStart \tab icon-start \tab Specifies an icon to display at the start of the component. \tab string \tab TRUE \cr
   value \tab value \tab The component's value. \tab any \tab FALSE \cr
}
}
\examples{
calcite_segmented_control_item()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/segmented-control-item/}{Official Documentation}
}
