% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_shell}
\alias{calcite_shell}
\title{Calcite Shell}
\usage{
calcite_shell(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-shell>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   contentBehind \tab content-behind \tab Positions the center content behind any calcite-shell-panels. \tab boolean \tab false \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding custom content. This content will appear between any leading and trailing panels added to the component, such as a map. \cr
   "alerts" \tab A slot for adding calcite-alert components. When placed in this slot, the alert position will be constrained to the extent of the calcite-shell. \cr
   "center-row" \tab DEPRECATED Use the "panel-bottom" slot instead. A slot for adding the bottom calcite-shell-center-row. \cr
   "dialogs" \tab A slot for adding calcite-dialog components. When placed in this slot, the dialog position will be constrained to the extent of the calcite-shell. \cr
   "footer" \tab A slot for adding footer content. This content will be positioned at the bottom of the component. \cr
   "header" \tab A slot for adding header content. This content will be positioned at the top of the component. \cr
   "modals" \tab A slot for adding calcite-modal components. When placed in this slot, the modal position will be constrained to the extent of the calcite-shell. \cr
   "panel-bottom" \tab A slot for adding the bottom calcite-shell-panel. \cr
   "panel-end" \tab A slot for adding the ending calcite-shell-panel. \cr
   "panel-start" \tab A slot for adding the starting calcite-shell-panel. \cr
   "panel-top" \tab A slot for adding the top calcite-shell-panel. \cr
   "sheets" \tab A slot for adding calcite-sheet components. When placed in this slot, the sheet position will be constrained to the extent of the calcite-shell. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/shell/}{Official Documentation}
}
