% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_slider}
\alias{calcite_slider}
\title{Calcite Slider}
\usage{
calcite_slider(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-slider>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   fillPlacement \tab fill-placement \tab Used to configure where the fill is placed along the slider track in relation to the value handle. Range mode will always display the fill between the min and max handles. \tab "end" | "none" | "start" \tab "start" \cr
   form \tab form \tab The id of the form that will be associated with the component. When not set, the component will be associated with its ancestor form element, if any. \tab string \tab undefined \cr
   groupSeparator \tab group-separator \tab When true, number values are displayed with a group separator corresponding to the language and country format. \tab boolean \tab false \cr
   hasHistogram \tab has-histogram \tab When true, indicates a histogram is present. \tab boolean \tab false \cr
   histogram \tab – \tab A list of the histogram’s x,y coordinates within the component’s min and max. Displays above the component’s track. \tab Point[] \tab undefined \cr
   histogramStops \tab – \tab A set of single color stops for a histogram, sorted by offset ascending. \tab ColorStop[] \tab undefined \cr
   labelFormatter \tab – \tab When specified, allows users to customize handle labels. \tab (value: number, type: "value" | "max" | "min" | "tick", defaultFormatter: (value: number) => string) => string \tab undefined \cr
   labelHandles \tab label-handles \tab When true, displays label handles with their numeric value. \tab boolean \tab false \cr
   labelTicks \tab label-ticks \tab When true and ticks is specified, displays label tick marks with their numeric value. \tab boolean \tab false \cr
   max \tab max \tab The component’s maximum selectable value. \tab number \tab 100 \cr
   maxLabel \tab max-label \tab For multiple selections, the accessible name for the second handle, such as "Temperature, upper bound". \tab string \tab undefined \cr
   maxValue \tab max-value \tab For multiple selections, the component’s upper value. \tab number \tab undefined \cr
   min \tab min \tab The component’s minimum selectable value. \tab number \tab 0 \cr
   minLabel \tab min-label \tab Accessible name for first (or only) handle, such as "Temperature, lower bound". \tab string \tab undefined \cr
   minValue \tab min-value \tab For multiple selections, the component’s lower value. \tab number \tab undefined \cr
   mirrored \tab mirrored \tab When true, the slider will display values from high to low. Note that this value will be ignored if the slider has an associated histogram. \tab boolean \tab false \cr
   name \tab name \tab Specifies the name of the component. Required to pass the component’s value on form submission. \tab string \tab undefined \cr
   numberingSystem \tab numbering-system \tab Specifies the Unicode numeral system used by the component for localization. \tab "arab" | "arabext" | "latn" \tab undefined \cr
   pageStep \tab page-step \tab Specifies the interval to move with the page up, or page down keys. \tab number \tab undefined \cr
   precise \tab precise \tab When true, sets a finer point for handles. \tab boolean \tab false \cr
   required \tab required \tab When true, the component must have a value in order for the form to submit. \tab boolean \tab false \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   snap \tab snap \tab When true, enables snap selection in coordination with step via a mouse. \tab boolean \tab false \cr
   step \tab step \tab Specifies the interval to move with the up, or down keys. \tab number \tab 1 \cr
   ticks \tab ticks \tab Displays tick marks on the number line at a specified interval. \tab number \tab undefined \cr
   value \tab value \tab The component’s value. \tab number | number[] \tab 0 \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteSliderChange \tab Fires when the thumb is released on the component. Note: To constantly listen to the drag event, use calciteSliderInput instead. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteSliderInput \tab Fires on all updates to the component. Note: Fires frequently during drag. To perform expensive operations consider using a debounce or throttle to avoid locking up the main thread. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/slider/}{Official Documentation}
}
