% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_sort_handle}
\alias{calcite_sort_handle}
\title{Calcite Sort Handle}
\usage{
calcite_sort_handle(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-sort-handle>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/sort-handle/}{Official Documentation}
}
