% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_sortable_list}
\alias{calcite_sortable_list}
\title{Calcite Sortable List}
\usage{
calcite_sortable_list(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-sortable-list>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   canPull \tab – \tab When provided, the method will be called to determine whether the element can move from the list. \tab (detail: DragDetail) => boolean \tab undefined \cr
   canPut \tab – \tab When provided, the method will be called to determine whether the element can be added from another list. \tab (detail: DragDetail) => boolean \tab undefined \cr
   disabled \tab disabled \tab When true, disabled prevents interaction. This state shows items with lower opacity/grayed. \tab boolean \tab false \cr
   dragSelector \tab drag-selector \tab Specifies which items inside the element should be draggable. \tab string \tab undefined \cr
   group \tab group \tab The list’s group identifier. To drag elements from one list into another, both lists must have the same group value. \tab string \tab undefined \cr
   handleSelector \tab handle-selector \tab The selector for the handle elements. \tab string \tab "calcite-handle" \cr
   layout \tab layout \tab Indicates the horizontal or vertical orientation of the component. \tab "grid" | "horizontal" | "vertical" \tab "vertical" \cr
   loading \tab loading \tab When true, content is waiting to be loaded. This state shows a busy indicator. \tab boolean \tab false \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteListOrderChange \tab Emitted when the order of the list has changed. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding sortable items. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/sortable-list/}{Official Documentation}
}
