% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_split_button}
\alias{calcite_split_button}
\title{Calcite Split Button}
\usage{
calcite_split_button(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-split-button>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   appearance \tab appearance \tab Specifies the appearance style of the component. \tab "outline" | "outline-fill" | "solid" | "transparent" \tab "solid" \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   dropdownIconType \tab dropdown-icon-type \tab Specifies the icon used for the dropdown menu. \tab "caret" | "chevron" | "ellipsis" | "overflow" \tab "chevron" \cr
   dropdownLabel \tab dropdown-label \tab Accessible name for the dropdown menu. \tab string \tab undefined \cr
   flipPlacements \tab – \tab Specifies the component’s fallback slotted content placement when it’s initial or specified placement has insufficient space available. \tab FlipPlacement[] \tab undefined \cr
   kind \tab kind \tab Specifies the kind of the component, which will apply to border and background, if applicable. \tab "brand" | "danger" | "inverse" | "neutral" \tab "brand" \cr
   loading \tab loading \tab When true, a busy indicator is displayed on the primary button. \tab boolean \tab false \cr
   overlayPositioning \tab overlay-positioning \tab Determines the type of positioning to use for the overlaid content. Using "absolute" will work for most cases. The component will be positioned inside of overflowing parent containers and will affect the container’s layout. "fixed" should be used to escape an overflowing parent container, or when the reference element’s position CSS property is "fixed". \tab "absolute" | "fixed" \tab "absolute" \cr
   placement \tab placement \tab Determines where the component will be positioned relative to the container element. \tab "bottom" | "bottom-end" | "bottom-start" | "top" | "top-end" | "top-start" \tab "bottom-end" \cr
   primaryIconEnd \tab primary-icon-end \tab Specifies an icon to display at the end of the primary button. \tab string \tab undefined \cr
   primaryIconFlipRtl \tab primary-icon-flip-rtl \tab Displays the primaryIconStart and/or primaryIconEnd as flipped when the element direction is right-to-left ("rtl"). \tab "both" | "end" | "start" \tab undefined \cr
   primaryIconStart \tab primary-icon-start \tab Specifies an icon to display at the start of the primary button. \tab string \tab undefined \cr
   primaryLabel \tab primary-label \tab Accessible name for the primary button. \tab string \tab undefined \cr
   primaryText \tab primary-text \tab Text displayed in the primary button. \tab string \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   width \tab width \tab Specifies the width of the component. \tab "auto" | "full" | "half" \tab "auto" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteSplitButtonPrimaryClick \tab Fires when the primary button is clicked. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteSplitButtonSecondaryClick \tab Fires when the dropdown menu is clicked. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-dropdown content. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/split-button/}{Official Documentation}
}
