% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_split_button}
\alias{calcite_split_button}
\title{Calcite Split Button}
\usage{
calcite_split_button(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-split-button>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/split-button/}{Official Documentation}
}
