% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_stack}
\alias{calcite_stack}
\title{Calcite Stack}
\usage{
calcite_stack(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-stack>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   disabled \tab disabled \tab When true, content interaction is prevented and displayed with lower opacity. \tab boolean \tab false \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding content. \cr
   "actions-end" \tab A slot for adding actionable calcite-action elements after the content of the component. \cr
   "actions-start" \tab A slot for adding actionable calcite-action elements before the content of the component. \cr
   "content-end" \tab A slot for adding non-actionable elements after content of the component. \cr
   "content-start" \tab A slot for adding non-actionable elements before content of the component. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/stack/}{Official Documentation}
}
