% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_stack}
\alias{calcite_stack}
\title{Create a Stack component}
\usage{
calcite_stack(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a Stack component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   disabled \tab disabled \tab When \code{true}, content interaction is prevented and displayed with lower opacity. \tab boolean \tab TRUE \cr
}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding content. \cr
   actions-start \tab A slot for adding actionable \code{calcite-action} elements before the content of the component. \cr
   content-start \tab A slot for adding non-actionable elements before content of the component. \cr
   content-end \tab A slot for adding non-actionable elements after content of the component. \cr
   actions-end \tab A slot for adding actionable \code{calcite-action} elements after the content of the component. \cr
}

}
}
\examples{
calcite_stack()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/stack/}{Official Documentation}
}
