% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_stepper}
\alias{calcite_stepper}
\title{Calcite Stepper}
\usage{
calcite_stepper(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-stepper>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   icon \tab icon \tab When true, displays a status icon in the calcite-stepper-item heading. \tab boolean \tab false \cr
   layout \tab layout \tab Defines the layout of the component. \tab "horizontal" | "horizontal-single" | "vertical" \tab "horizontal" \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab StepperMessages \tab undefined \cr
   numbered \tab numbered \tab When true, displays the step number in the calcite-stepper-item heading. \tab boolean \tab false \cr
   numberingSystem \tab numbering-system \tab Specifies the Unicode numeral system used by the component for localization. \tab "arab" | "arabext" | "latn" \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   selectedItem \tab – \tab Specifies the component’s selected item. \tab HTMLCalciteStepperItemElement \tab null \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteStepperChange \tab Fires when the active calcite-stepper-item changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
   calciteStepperItemChange \tab DEPRECATED use calciteStepperChange instead or calciteStepperItemChange on items instead.Fires when the active calcite-stepper-item changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-stepper-item elements. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/stepper/}{Official Documentation}
}
