% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_stepper_item}
\alias{calcite_stepper_item}
\title{Calcite Stepper Item}
\usage{
calcite_stepper_item(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-stepper-item>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   complete \tab complete \tab When true, the step has been completed. \tab boolean \tab false \cr
   description \tab description \tab A description for the component. Displays below the header text. \tab string \tab undefined \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   error \tab error \tab When true, the component contains an error that requires resolution from the user. \tab boolean \tab false \cr
   heading \tab heading \tab The component header text. \tab string \tab undefined \cr
   iconFlipRtl \tab icon-flip-rtl \tab When true, the icon will be flipped when the element direction is right-to-left ("rtl"). \tab boolean \tab false \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab StepperItemMessages \tab undefined \cr
   selected \tab selected \tab When true, the component is selected. \tab boolean \tab false \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteStepperItemSelect \tab Fires when the active calcite-stepper-item changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding custom content. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/stepper-item/}{Official Documentation}
}
