% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_switch}
\alias{calcite_switch}
\title{Calcite Switch}
\usage{
calcite_switch(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-switch>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/switch/}{Official Documentation}
}
