% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tab_title}
\alias{calcite_tab_title}
\title{Calcite Tab Title}
\usage{
calcite_tab_title(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tab-title>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tab-title/}{Official Documentation}
}
