% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_table}
\alias{calcite_table}
\title{Calcite Table}
\usage{
calcite_table(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-table>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/table/}{Official Documentation}
}
