% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_table_cell}
\alias{calcite_table_cell}
\title{Calcite Table Cell}
\usage{
calcite_table_cell(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-table-cell>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/table-cell/}{Official Documentation}
}
