% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_table_cell}
\alias{calcite_table_cell}
\title{Create a TableCell component}
\usage{
calcite_table_cell(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a TableCell component
}
\details{
#' ## Properties
The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   alignment \tab alignment \tab Specifies the alignment of the component. \tab "center" | "end" | "start" \tab TRUE \cr
   colSpan \tab col-span \tab Specifies the number of columns the component should span. \tab number \tab TRUE \cr
   messageOverrides \tab NA \tab Use this property to override individual strings used by the component. \tab Check API reference \tab FALSE \cr
   rowSpan \tab row-span \tab Specifies the number of rows the component should span. \tab number \tab TRUE \cr
}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding content, usually text content. \cr
}

}
}
\examples{
calcite_table_cell()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/table-cell/}{Official Documentation}
}
