% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_table_header}
\alias{calcite_table_header}
\title{Calcite Table Header}
\usage{
calcite_table_header(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-table-header>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   alignment \tab alignment \tab Specifies the alignment of the component. \tab "center" | "end" | "start" \tab "start" \cr
   colSpan \tab col-span \tab Specifies the number of columns the component should span. \tab number \tab undefined \cr
   description \tab description \tab A description to display beneath heading content. \tab string \tab undefined \cr
   heading \tab heading \tab A heading to display above description content. \tab string \tab undefined \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab TableHeaderMessages \tab undefined \cr
   rowSpan \tab row-span \tab Specifies the number of rows the component should span. \tab number \tab undefined \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/table-header/}{Official Documentation}
}
