% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_table_row}
\alias{calcite_table_row}
\title{Calcite Table Row}
\usage{
calcite_table_row(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-table-row>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   alignment \tab alignment \tab Specifies the alignment of the component. \tab "center" | "end" | "start" \tab undefined \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   selected \tab selected \tab When true, the component is selected. \tab boolean \tab false \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteTableRowSelect \tab Fires when the selected state of the component changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-table-cell or calcite-table-header elements. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/table-row/}{Official Documentation}
}
