% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_text_area}
\alias{calcite_text_area}
\title{Calcite Text Area}
\usage{
calcite_text_area(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-text-area>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/text-area/}{Official Documentation}
}
