% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tile_group}
\alias{calcite_tile_group}
\title{Calcite Tile Group}
\usage{
calcite_tile_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tile-group>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   alignment \tab alignment \tab Specifies the alignment of each calcite-tile’s content. \tab "center" | "start" \tab "start" \cr
   disabled \tab disabled \tab When true, interaction is prevented and the component is displayed with lower opacity. \tab boolean \tab false \cr
   label (required) \tab label \tab Accessible name for the component. \tab string \tab undefined \cr
   layout \tab layout \tab Defines the layout of the component. Use "horizontal" for rows, and "vertical" for a single column. \tab "horizontal" | "vertical" \tab "horizontal" \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   selectedItems \tab – \tab Specifies the component’s selected items. \tab HTMLCalciteTileElement[] \tab [] \cr
   selectionAppearance \tab selection-appearance \tab Specifies the selection appearance, where: - "icon" (displays a checkmark or dot), or - "border" (displays a border). \tab "border" | "icon" \tab "icon" \cr
   selectionMode \tab selection-mode \tab Specifies the selection mode, where: - "multiple" (allows any number of selected items), - "single" (allows only one selected item), - "single-persist" (allows only one selected item and prevents de-selection), - "none" (allows no selected items). \tab "multiple" | "none" | "single" | "single-persist" \tab "none" \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteTileGroupSelect \tab Fires when the component’s selection changes. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
   NA \tab A slot for adding calcite-tile elements. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tile-group/}{Official Documentation}
}
