% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tile_select_group}
\alias{calcite_tile_select_group}
\title{Calcite Tile Select Group}
\usage{
calcite_tile_select_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tile-select-group>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tile-select-group/}{Official Documentation}
}
