% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_time_picker}
\alias{calcite_time_picker}
\title{Calcite Time Picker}
\usage{
calcite_time_picker(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-time-picker>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab TimePickerMessages \tab undefined \cr
   numberingSystem \tab numbering-system \tab Specifies the Unicode numeral system used by the component for localization. \tab "arab" | "arabext" | "latn" \tab undefined \cr
   scale \tab scale \tab Specifies the size of the component. \tab "l" | "m" | "s" \tab "m" \cr
   step \tab step \tab Specifies the granularity the value must adhere to (in seconds). \tab number \tab 60 \cr
   value \tab value \tab The component’s value in UTC (always 24-hour format). \tab string \tab null \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/time-picker/}{Official Documentation}
}
