% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_time_picker}
\alias{calcite_time_picker}
\title{Calcite Time Picker}
\usage{
calcite_time_picker(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-time-picker>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/time-picker/}{Official Documentation}
}
