% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tip}
\alias{calcite_tip}
\title{Calcite Tip}
\usage{
calcite_tip(...)
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tip>} component.
}
\section{Component Details}{


The below tables illustrate the properties, attributes, and optionally the
slots, events, and methods associated with this component.

Each attribute can be set through the corresponding argument in this function
or by passing it in via \code{...} as a named value e.g. \code{"icon-start" = "snow"}.

Each event is subscribed to by shiny. The component properties are updated
on each of the events.
\subsection{Properties}{\tabular{lllll}{
   Property \tab Attribute \tab Description \tab Type \tab Default \cr
   closeDisabled \tab close-disabled \tab When true, the close button is not present on the component. \tab boolean \tab false \cr
   closed \tab closed \tab When true, the component does not display. \tab boolean \tab false \cr
   heading \tab heading \tab The component header text. \tab string \tab undefined \cr
   headingLevel \tab heading-level \tab Specifies the heading level of the component’s heading for proper document structure, without affecting visual styling. \tab 1 | 2 | 3 | 4 | 5 | 6 \tab undefined \cr
   messageOverrides \tab message-overrides \tab Use this property to override individual strings used by the component. \tab TipMessages \tab undefined \cr
   selected \tab selected \tab When true, the component is selected if it has a parent calcite-tip-manager. Only one tip can be selected within the calcite-tip-manager parent. \tab boolean \tab false \cr
}

}

\subsection{Events}{\tabular{lll}{
   Event \tab Description \tab Type \cr
   calciteTipDismiss \tab Emits when the component has been closed. \tab CustomEvent\if{html}{\out{<void>}} \cr
}

}

\subsection{Slots}{\tabular{ll}{
   Slot \tab Description \cr
    \tab A slot for adding text and a hyperlink. \cr
   "thumbnail" \tab A slot for adding an HTML image element. \cr
}

}
}

\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tip/}{Official Documentation}
}
