% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tip_group}
\alias{calcite_tip_group}
\title{Calcite Tip Group}
\usage{
calcite_tip_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Create a \verb{<calcite-tip-group>} component.
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tip-group/}{Official Documentation}
}
