% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-generated.R
\name{calcite_tip_group}
\alias{calcite_tip_group}
\title{Create a TipGroup component}
\usage{
calcite_tip_group(...)
}
\arguments{
\item{...}{named attributes passed to \code{htmltools::tag()}}
}
\value{
an object of class \code{calcite_component} which is a subclass of \code{shiny.tag}
}
\description{
Use the \code{calcite-carousel} and \code{calcite-carousel-item} components instead.
}
\details{
\subsection{Properties}{

The following properties are provided by this component:\tabular{lllll}{
   Name \tab Attribute \tab Description \tab Values \tab Reflects to Attribute \cr
   groupTitle \tab group-title \tab The component header text for all nested \code{calcite-tip}s. \tab string \tab FALSE \cr
}

}

\subsection{Slots}{

The following slots are provided by this component:\tabular{ll}{
   Slot \tab Description \cr
   Default (unnamed) \tab A slot for adding \code{calcite-tip}s. \cr
}

}
}
\examples{
calcite_tip_group()
}
\references{
\href{https://developers.arcgis.com/calcite-design-system/components/tip-group/}{Official Documentation}
}
